+++ variabledynarr.h
@@ -2,8 +2,8 @@
 #include "variable.h"
 #include "dynarr.h"
 
-void DynArrVar_print(void* data, int len) {
-	for (int i = 0; i < len; i++) {
+void DynArrVar_print(void* data, size_t len) {
+	for (size_t i = 0; i < len; i++) {
 		Variable_print(data + i, true);
 	}
 }
